refactor(networks): signer.ts hold the signer code and index.ts still holds network utils
