+++ packages/hooks/dao/useDAOMakeProposal.ts
@@ -6,11 +6,8 @@ import { useInvalidateDAOProposals } from "./useDAOProposals";
 import { DaoPreProposeSingleClient } from "@/contracts-clients/dao-pre-propose-single/DaoPreProposeSingle.client";
 import { CosmosMsgForEmpty } from "@/contracts-clients/dao-pre-propose-single/DaoPreProposeSingle.types";
 import { DaoProposalSingleQueryClient } from "@/contracts-clients/dao-proposal-single/DaoProposalSingle.client";
-import {
-  mustGetNonSigningCosmWasmClient,
-  getKeplrSigningCosmWasmClient,
-  parseUserId,
-} from "@/networks";
+import { mustGetNonSigningCosmWasmClient, parseUserId } from "@/networks";
+import { getKeplrSigningCosmWasmClient } from "@/networks/signer";
 
 export const useDAOMakeProposal = (
   daoId: string | undefined,
