+++ packages/networks/index.ts
@@ -1,24 +1,8 @@
-import {
-  CosmWasmClient,
-  SigningCosmWasmClient,
-} from "@cosmjs/cosmwasm-stargate";
-import {
-  createWasmAminoConverters,
-  wasmTypes,
-} from "@cosmjs/cosmwasm-stargate/build/modules";
+import { CosmWasmClient } from "@cosmjs/cosmwasm-stargate";
 import { Decimal } from "@cosmjs/math";
-import { Registry } from "@cosmjs/proto-signing";
-import {
-  AminoTypes,
-  createDefaultAminoConverters,
-  defaultRegistryTypes,
-  GasPrice,
-  SigningStargateClient,
-  StargateClient,
-} from "@cosmjs/stargate";
+import { GasPrice, StargateClient } from "@cosmjs/stargate";
 import { ChainInfo, Currency as KeplrCurrency } from "@keplr-wallet/types";
 import { bech32 } from "bech32";
-import { Platform } from "react-native";
 
 import { cosmosNetwork } from "./cosmos-hub";
 import { cosmosThetaNetwork } from "./cosmos-hub-theta";
@@ -45,13 +29,6 @@ import {
   NetworkInfo,
   NetworkKind,
 } from "./types";
-import {
-  teritoriAminoConverters,
-  teritoriProtoRegistry,
-} from "../api/teritori-chain";
-import { convertKeplrSigner, getKeplr } from "../utils/keplr";
-
-import { getNativeWallet } from "@/utils/wallet/getNativeWallet";
 
 export * from "./types";
 
@@ -95,18 +72,6 @@ export const allNetworks = [
     (rn) => !packageNetworks.some((pn) => pn.overrides === rn.id),
   ),
 ].sort((a, b) => a.displayName.localeCompare(b.displayName));
-
-export const cosmosTypesRegistry = new Registry([
-  ...defaultRegistryTypes,
-  ...wasmTypes,
-  ...teritoriProtoRegistry,
-]);
-
-const cosmosAminoTypes = new AminoTypes({
-  ...createDefaultAminoConverters(),
-  ...createWasmAminoConverters(),
-  ...teritoriAminoConverters,
-});
 
 export const getCurrency = (
   networkId: string | undefined,
@@ -446,56 +411,6 @@ export const cosmosNetworkGasPrice = (
   return new GasPrice(decimalGasPrice, feeCurrency.denom);
 };
 
-export const getKeplrSigner = async (networkId: string) => {
-  const network = mustGetCosmosNetwork(networkId);
-
-  const keplr = getKeplr();
-
-  await keplr.experimentalSuggestChain(keplrChainInfoFromNetworkInfo(network));
-
-  await keplr.enable(network.chainId);
-
-  const keplrSigner = await keplr.getOfflineSignerAuto(network.chainId);
-
-  return convertKeplrSigner(keplrSigner);
-};
-
-const getKeplrOnlyAminoSigner = async (networkId: string) => {
-  const network = mustGetCosmosNetwork(networkId);
-
-  const keplr = getKeplr();
-
-  await keplr.experimentalSuggestChain(keplrChainInfoFromNetworkInfo(network));
-
-  await keplr.enable(network.chainId);
-
-  return keplr.getOfflineSignerOnlyAmino(network.chainId);
-};
-
-export const getKeplrSigningStargateClient = async (
-  networkId: string,
-  gasPriceKind: "low" | "average" | "high" = "average",
-) => {
-  const network = mustGetCosmosNetwork(networkId);
-
-  const gasPrice = cosmosNetworkGasPrice(network, gasPriceKind);
-  if (!gasPrice) {
-    throw new Error("gas price not found");
-  }
-
-  const signer = await getKeplrSigner(networkId);
-
-  return await SigningStargateClient.connectWithSigner(
-    network.rpcEndpoint,
-    signer,
-    {
-      gasPrice,
-      registry: cosmosTypesRegistry,
-      aminoTypes: cosmosAminoTypes,
-    },
-  );
-};
-
 export const getNetworkFeature = <
   F extends NetworkFeature,
   FO extends NetworkFeatureObject,
@@ -514,58 +429,6 @@ export const getNetworkFeature = <
   return network.featureObjects.find((f) => f.type === feature) as
     | R
     | undefined;
-};
-
-export const getKeplrOnlyAminoStargateClient = async (
-  networkId: string,
-  gasPriceKind: "low" | "average" | "high" = "average",
-) => {
-  const network = mustGetCosmosNetwork(networkId);
-
-  const gasPrice = cosmosNetworkGasPrice(network, gasPriceKind);
-  if (!gasPrice) {
-    throw new Error("gas price not found");
-  }
-
-  const signer = await getKeplrOnlyAminoSigner(networkId);
-
-  return await SigningStargateClient.connectWithSigner(
-    network.rpcEndpoint,
-    signer,
-    {
-      gasPrice,
-      registry: cosmosTypesRegistry,
-      aminoTypes: cosmosAminoTypes,
-    },
-  );
-};
-
-export const getKeplrSigningCosmWasmClient = async (
-  networkId: string,
-  gasPriceKind: "low" | "average" | "high" = "average",
-) => {
-  const network = mustGetCosmosNetwork(networkId);
-  const gasPrice = cosmosNetworkGasPrice(network, gasPriceKind);
-  if (!gasPrice) {
-    throw new Error("gas price not found");
-  }
-  if (Platform.OS !== "web") {
-    const wallet = await getNativeWallet(network.addressPrefix, 1); // todo make multi wallet
-
-    return SigningCosmWasmClient.connectWithSigner(
-      network.rpcEndpoint,
-      wallet,
-      {
-        gasPrice,
-      },
-    );
-  }
-
-  const signer = await getKeplrSigner(networkId);
-
-  return SigningCosmWasmClient.connectWithSigner(network.rpcEndpoint, signer, {
-    gasPrice,
-  });
 };
 
 export const mustGetNonSigningCosmWasmClient = async (networkId: string) => {
